+++ components/openapi/src/main/scala/pl/touk/nussknacker/openapi/extractor/ParametersExtractor.scala
@@ -10,7 +10,7 @@ object ParametersExtractor {
   def queryParams(paramDef: QueryParameter, paramInput: Any): List[(String, String)] = {
     import scala.collection.JavaConverters._
     paramDef.`type` match {
-      case SwaggerObject(fieldDefs) =>
+      case SwaggerObject(fieldDefs, _) =>
         val inputs = paramInput.asInstanceOf[java.util.Map[String, AnyRef]].asScala
         inputs.toList.flatMap { case (a, b) =>
           queryParams(QueryParameter(s"${paramDef.name}.$a", fieldDefs(a)), b)
@@ -27,7 +27,7 @@ object ParametersExtractor {
 
   private def flattenBodyParameter(bodyParameter: SingleBodyParameter): List[ParameterWithBodyFlag] = {
     bodyParameter.`type` match {
-      case SwaggerObject(elementType) =>
+      case SwaggerObject(elementType, _) =>
         elementType.map { case (propertyName, swaggerType) =>
           prepareParameter(propertyName, swaggerType, isBodyPart = true)
         }.toList