+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/serde/CirceJsonDeserializer.scala
@@ -28,7 +28,8 @@ class CirceJsonDeserializer(jsonSchema: Schema) {
     val jsonObject = new JSONTokener(string).nextValue()
     //after validate jsonObject has set default field values
     catchValidationError(jsonSchema.validate(jsonObject))
-    JsonToNuStruct(toCirce(jsonObject), swaggerTyped)
+    val struct = JsonToNuStruct(toCirce(jsonObject), swaggerTyped)
+    struct
   }
 
   private def toCirce(json: Object): Json = json match {